+++ packages/pangea-sdk/tests/integration/asyncio/__init__.py
@@ -1,5 +1,6 @@
 from .test_audit import TestAuditAsync  # noqa: F401
 from .test_authn import TestAuthN  # noqa: F401
+from .test_authz import TestAuthZIntegration  # noqa: F401
 from .test_embargo import TestEmbargo  # noqa: F401
 from .test_file_scan import TestFileScan  # noqa: F401
 from .test_intel import TestDomainIntel, TestFileIntel, TestIPIntel, TestURLIntel, TestUserIntel  # noqa: F401