+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for Secure Audit Log's log stream API.
 - Support for Secure Audit Log's export API.
+- AuthZ service support.
 
 ### Fixed